# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on mutating money endpoints. A UUID the client generates per intent.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Email or phone already registered."
          }
        },
        "summary": "Register a client or promoter",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/otp/request": {
      "post": {
        "description": "Always returns 202, whether or not the number is registered — an unauthenticated caller does not get to enumerate accounts.",
        "operationId": "AuthController_requestOtp",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on mutating money endpoints. A UUID the client generates per intent.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptedDto"
                }
              }
            }
          }
        },
        "summary": "Request a verification code",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/otp/verify": {
      "post": {
        "operationId": "AuthController_verifyOtp",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on mutating money endpoints. A UUID the client generates per intent.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          }
        },
        "summary": "Verify a phone code and receive tokens",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on mutating money endpoints. A UUID the client generates per intent.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          },
          "401": {
            "description": "Email or password is incorrect."
          },
          "403": {
            "description": "Account suspended, or phone not yet verified (code PHONE_NOT_VERIFIED)."
          }
        },
        "summary": "Log in with email and password",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "description": "Rotates: the presented token is revoked. Replaying an already-revoked token revokes every session for that user.",
        "operationId": "AuthController_refresh",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on mutating money endpoints. A UUID the client generates per intent.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairDto"
                }
              }
            }
          }
        },
        "summary": "Exchange a refresh token for a new pair",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on mutating money endpoints. A UUID the client generates per intent.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Revoke a refresh token",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on mutating money endpoints. A UUID the client generates per intent.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "The current user",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/change-password": {
      "post": {
        "description": "Requires the current password. Revokes all other sessions on success.",
        "operationId": "AuthController_changePassword",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on mutating money endpoints. A UUID the client generates per intent.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Change password",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ralia API",
    "description": "Two-sided promotion marketplace — businesses fund campaigns, promoters post them for a fee.\n\n**Money**: every amount is an integer in minor units (kobo). Responses carry both\n`amount_minor` (integer) and `amount_display` (formatted). Never a float.\n\n**Idempotency**: every mutating money endpoint requires an `Idempotency-Key` header\nand rejects the request without one. Replaying a key returns the original result\nwithout moving the balance again.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ada@example.com"
          },
          "phone_e164": {
            "type": "string",
            "example": "+2348012345678",
            "description": "E.164, with country code."
          },
          "password": {
            "type": "string",
            "example": "correct horse battery staple",
            "minLength": 10
          },
          "role": {
            "type": "string",
            "enum": [
              "CLIENT",
              "PROMOTER"
            ],
            "example": "PROMOTER"
          },
          "org_name": {
            "type": "string",
            "example": "Naija Threads",
            "description": "Required when role is CLIENT."
          },
          "accepted_terms": {
            "type": "boolean",
            "example": true,
            "description": "Must be true. Recorded as a consent row."
          },
          "accepted_privacy": {
            "type": "boolean",
            "example": true,
            "description": "Must be true. Recorded as a consent row."
          }
        },
        "required": [
          "email",
          "phone_e164",
          "password",
          "role",
          "accepted_terms",
          "accepted_privacy"
        ]
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "example": "PENDING"
          },
          "next": {
            "type": "string",
            "example": "VERIFY_PHONE",
            "description": "What the client should do next."
          }
        },
        "required": [
          "user_id",
          "status",
          "next"
        ]
      },
      "OtpRequestDto": {
        "type": "object",
        "properties": {
          "phone_e164": {
            "type": "string",
            "example": "+2348012345678"
          }
        },
        "required": [
          "phone_e164"
        ]
      },
      "AcceptedDto": {
        "type": "object",
        "properties": {
          "accepted": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "If that number is registered, a code has been sent."
          }
        },
        "required": [
          "accepted",
          "message"
        ]
      },
      "OtpVerifyDto": {
        "type": "object",
        "properties": {
          "phone_e164": {
            "type": "string",
            "example": "+2348012345678"
          },
          "code": {
            "type": "string",
            "example": "123456",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "phone_e164",
          "code"
        ]
      },
      "TokenPairDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "expires_in": {
            "type": "number",
            "example": 900,
            "description": "Access token lifetime in seconds."
          },
          "token_type": {
            "type": "string",
            "example": "Bearer"
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "token_type"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "ada@example.com"
          },
          "password": {
            "type": "string",
            "example": "correct horse battery staple"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        },
        "required": [
          "refresh_token"
        ]
      },
      "MeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string"
          },
          "phone_e164": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "CLIENT",
                "PROMOTER",
                "ADMIN"
              ]
            }
          },
          "status": {
            "type": "string"
          },
          "phone_verified_at": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "email",
          "phone_e164",
          "roles",
          "status",
          "phone_verified_at"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "current_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string",
            "minLength": 10
          }
        },
        "required": [
          "current_password",
          "new_password"
        ]
      },
...
    }
  }
}